Split PersonaCoin
