+++ packages/office-ui-fabric-react/src/components/Persona/index.ts
@@ -2,4 +2,5 @@ export * from './Persona';
 export * from './Persona.base';
 export * from './Persona.types';
 export * from './PersonaCoin';
+export * from './PersonaCoin.base';
 export * from './PersonaConsts';
